+++ pyechelle/CCD.py
@@ -1,56 +1,9 @@
 import logging
-from concurrent.futures import ThreadPoolExecutor
-from functools import partial
 
 import h5py
 import numpy as np
-from numba import int32, float64, int64, njit
 
 logger = logging.getLogger('CCD')
-
-
-@njit(
-    int32[:, :](float64[:], float64[:], int64, int64, int64, int64),
-    parallel=True,
-    nogil=True,
-    cache=True,
-)
-def bin_2d(x, y, xmin=0, xmax=4096, ymin=0, ymax=4096):
-    """  Bin XY position into 2d grid and throw away data outside the limits.
-
-    Args:
-        x (np.ndarray): X positions
-        y (np.ndarray): Y positions
-        xmin (int): minimum X value of the grid
-        xmax (int): maximum X value of the grid
-        ymin (int): minimum Y value of the grid
-        ymax (int): maximum Y value of the grid
-
-    Returns:
-        np.ndarray: binned XY positions
-    """
-    grid = np.zeros((ymax - ymin, xmax - xmin), dtype=np.int32)
-
-    valid_idx = np.logical_and(x >= xmin, y >= ymin)
-    valid_idx = np.logical_and(valid_idx, y < ymax)
-    valid_idx = np.logical_and(valid_idx, x < xmax)
-    x = x[valid_idx]
-    y = y[valid_idx]
-    for xx, yy in zip(x, y):
-        grid[int(yy), int(xx)] += 1
-    return grid
-
-
-def _bin_2d(x, y, xmin=0, xmax=4096, ymin=0, ymax=4096):
-    splits = 24
-    with ThreadPoolExecutor(max_workers=splits) as pool:
-        chunk = x.shape[0] // splits
-        chunksx = [x[i * chunk:(i + 1) * chunk] for i in range(splits)]
-        chunksy = [y[i * chunk:(i + 1) * chunk] for i in range(splits)]
-        f = partial(_bin_2d, xmin=xmin, xmax=xmax, ymin=ymin, ymax=ymax)
-        results = pool.map(f, chunksx, chunksy)
-        results = np.sum(results)
-    return results
 
 
 def read_ccd_from_hdf(path):
@@ -64,7 +17,7 @@ def read_ccd_from_hdf(path):
 
 class CCD:
     def __init__(self, name='detector', xmin=0, xmax=4096, ymin=0, ymax=4096, maxval=65536, pixelsize=9):
-        self.data = np.zeros(((ymax - ymin), (xmax - xmin)), dtype=np.int32)
+        self.data = np.zeros(((ymax - ymin), (xmax - xmin)), dtype=np.uint32)
         self.name = name
         self.xmin = xmin
         self.xmax = xmax
@@ -73,11 +26,8 @@ class CCD:
         self.maxval = maxval
         self.pixelsize = pixelsize
 
-    def add_photons(self, x_positions, y_positions):
-        self.data += bin_2d(x_positions, y_positions, self.xmin, self.xmax, self.ymin, self.ymax)
-
     def add_readnoise(self, std=3.):
-        self.data += np.asarray(np.random.normal(0., std, self.data.shape).round(0), dtype=np.int32)
+        self.data = self.data + np.asarray(np.random.normal(0., std, self.data.shape).round(0), dtype=np.int32)
 
     def add_bias(self, value: int = 1000):
         """Adds a bias value to the detector counts