+++ CHANGELOG.md
@@ -9,13 +9,14 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
+- CUDA support (use the --cuda flag)
 - more tests to improve code reliability and increase coverage
 
 ### Changed
 
 - restructured core algorithm, code now properly parallelized by order
 - better scaling with number of CPUs for multi-order simulations
-- speed improvements
+- speed improvements for CPU simulation
 - more precise simulation by interpolating transformation matrices
 
 ### Removed